+++ src/components/scenes/SceneEntryHeader.jsx
@@ -1,7 +1,11 @@
 import { useRef } from 'react';
 
-import { useSceneHeading, useSceneNumber } from '@/stores/document';
-import { useSceneShotCount } from '@/stores/document/use';
+import {
+  isSceneEmpty,
+  useSceneHeading,
+  useSceneNumber,
+} from '@/stores/document';
+import { useDocumentStore } from '@/stores/document/use';
 import { useCurrentCursor, useSetUserCursor } from '@/stores/user';
 import BarberpoleStyle from '@/styles/Barberpole.module.css';
 
@@ -17,7 +21,9 @@ export default function SceneEntryHeader({ className, documentId, sceneId }) {
   const containerRef = useRef(/** @type {HTMLDivElement|null} */ (null));
   const [sceneHeading, setSceneHeading] = useSceneHeading(documentId, sceneId);
   const sceneNumber = useSceneNumber(documentId, sceneId);
-  const sceneShotCount = useSceneShotCount(documentId, sceneId);
+  const sceneEmpty = useDocumentStore((ctx) =>
+    isSceneEmpty(ctx, documentId, sceneId),
+  );
   const currentCursor = useCurrentCursor();
   const setUserCursor = useSetUserCursor();
   const isActive =
@@ -63,7 +69,7 @@ export default function SceneEntryHeader({ className, documentId, sceneId }) {
         className={
           'flex-1 w-full bg-transparent px-2 py-6 text-xl font-bold' +
           ' ' +
-          (sceneShotCount <= 0 ? 'line-through' : '')
+          (sceneEmpty ? 'line-through' : '')
         }
         type="text"
         list="sceneHeading"